Introduce PeopleRecommendationCommander in Shoebox
